Remove HeatSource.hh
