+++ source/CubeHeatSource.cc
@@ -15,7 +15,6 @@ namespace adamantine
 {
 template <int dim>
 CubeHeatSource<dim>::CubeHeatSource(boost::property_tree::ptree const &database)
-    : HeatSource<dim, dealii::MemorySpace::Default>()
 {
   _start_time = database.get<double>("start_time");
   _end_time = database.get<double>("end_time");
