+++ source/material_deposition.cc
@@ -6,6 +6,7 @@
  */
 
 #include <material_deposition.hh>
+#include <types.hh>
 #include <utils.hh>
 
 #include <deal.II/arborx/bvh.h>
@@ -400,10 +401,12 @@ merge_deposition_paths(
 
 template std::tuple<std::vector<dealii::BoundingBox<2>>, std::vector<double>,
                     std::vector<double>, std::vector<double>>
-deposition_along_scan_path(boost::property_tree::ptree const &geometry_database,
-                           ScanPath const &scan_path);
+deposition_along_scan_path(
+    boost::property_tree::ptree const &geometry_database,
+    ScanPath<dealii::MemorySpace::Host> const &scan_path);
 template std::tuple<std::vector<dealii::BoundingBox<3>>, std::vector<double>,
                     std::vector<double>, std::vector<double>>
-deposition_along_scan_path(boost::property_tree::ptree const &geometry_database,
-                           ScanPath const &scan_path);
+deposition_along_scan_path(
+    boost::property_tree::ptree const &geometry_database,
+    ScanPath<dealii::MemorySpace::Host> const &scan_path);
 } // namespace adamantine
